final dataset gen script
